macOS: move NSRange.clamped(toTextLength:) out of the attachment reconciler
Both draft stores use it for selection clamping, and neither has anything to do with
attachment reconciliation — it was just sitting at the bottom of that file. Lands in
Common/Extensions alongside the other NS* extensions. No behaviour change.

Reported by tomasstrba on #6064.
