+++ macOS/DuckDuckGo/Common/Extensions/NSRangeExtension.swift
@@ -1,0 +1,34 @@
+//
+//  NSRangeExtension.swift
+//
+//  Copyright © 2026 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+
+extension NSRange {
+
+    /// Keeps a stored selection valid against text that shrank under it: a start past the end
+    /// collapses to a caret there, an overhanging end truncates.
+    func clamped(toTextLength length: Int) -> NSRange {
+        if location > length {
+            return NSRange(location: length, length: 0)
+        }
+        if upperBound > length {
+            return NSRange(location: location, length: max(0, length - location))
+        }
+        return self
+    }
+}
